requirements.txt
certifi==2018.1.18
chardet==3.0.4
coverage==4.5.1
dj-database-url==0.4.2
Django==2.0.3
django-webpack-loader==0.6.0
djangorestframework==3.7.7
docopt==0.6.2
gunicorn==19.7.1
idna==2.6
pipreqs==0.4.9
psycopg2==2.7.4
pytz==2018.3
requests==2.18.4
six==1.11.0
typing==3.6.4
urllib3==1.22
whitenoise==3.3.1
yarg==0.1.9
